+++ frontend/private-plugins/block_ui/src/modules/status/graphql/mutations/updateBlockStatusOrder.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const UPDATE_BLOCK_STATUS_ORDER = gql`
-  mutation updateBlockOpptyStatusOrder($_id: String!, $order: Float!) {
+  mutation updateBlockOpptyStatusOrder($_id: String!, $order: Int!) {
     updateBlockOpptyStatusOrder(_id: $_id, order: $order) {
       _id
       order
